apps/web-app/tests/pages/api/members/[id].spec.ts
import airtableService from '@protocol-labs-network/airtable';
import { IMember } from '@protocol-labs-network/api';
import { NextApiRequest, NextApiResponse } from 'next';
import getMemberHandler from '../../../../pages/api/members/[id]';

const mockMember = {} as IMember;
jest.mock('@protocol-labs-network/airtable', () => ({
  getMember: jest.fn(() => mockMember),
}));

describe('/api/members/:id', () => {
  const id = 'memberId_01';
  const json = jest.fn();
  const end = jest.fn();
  const setHeader = jest.fn();
  const status = jest.fn(() => ({ json, end }));
  const req: NextApiRequest = {
    method: 'GET',
    query: { id },
  } as unknown as NextApiRequest;
  const res: NextApiResponse = {
    status,
    setHeader,
  } as unknown as NextApiResponse;

  beforeEach(() => {
    jest.clearAllMocks();
    getMemberHandler(req, res);
  });

  describe('and the request method is not supported', () => {
    let oldMethod: string;

    beforeAll(() => {
      oldMethod = req.method;
      req.method = 'POST';
    });

    it('should respond with an error', () => {
      expect(setHeader).toHaveBeenCalledTimes(1);
      expect(setHeader).toHaveBeenCalledWith('Allow', ['GET']);
      expect(status).toHaveBeenCalledTimes(1);
      expect(status).toHaveBeenCalledWith(405);
      expect(end).toHaveBeenCalledTimes(1);
      expect(end).toHaveBeenCalledWith('🚫 Method POST Not Allowed');
    });

    afterAll(() => {
      req.method = oldMethod;
    });
  });

  it('should get the member with the provided id from AirtableService', () => {
    expect(airtableService.getMember).toHaveBeenCalledTimes(1);
    expect(airtableService.getMember).toHaveBeenCalledWith(id);
  });

  describe('and AirtableService successfully retrieves a member', () => {
    it('should respond with the provided member', () => {
      expect(status).toHaveBeenCalledTimes(1);
      expect(status).toHaveBeenCalledWith(200);
      expect(json).toHaveBeenCalledTimes(1);
      expect(json).toHaveBeenCalledWith(mockMember);
    });
  });

  describe('and AirtableService fails to retrieve a member', () => {
    beforeAll(() => {
      jest.spyOn(airtableService, 'getMember').mockImplementationOnce(() => {
        throw new Error();
      });
    });

    it('should respond with an error', () => {
      expect(status).toHaveBeenCalledTimes(1);
      expect(status).toHaveBeenCalledWith(500);
      expect(json).toHaveBeenCalledTimes(1);
      expect(json).toHaveBeenCalledWith({
        msg: 'Ups, something went wrong 😕',
      });
    });
  });
});
